# schemas/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: We Sell Houses API Documentation
  description: API for property management and communication between user and property agent
  contact:
    name: Filip Prosovsky
    email: prosovskyf@uni.coventry.ac.uk
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: 'https://observe-prime.codio-box.uk/api/v1'
    description: Development server for this application api
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    BearerAuth:
      type: http
      scheme: bearer
...
paths:
...
    /admin/properties/data:
      summary: Zoopla data
      description: >
        Get zoopla data from database if they were already fetched before or fetch 
        new ones which are not yet stored in the database
      post:
        tags:
          - Zoopla
        description:
          Get Zoopla data based on Postcode
        security:
          - BearerAuth: []
        requestBody:
          description: Postcode
          required: true
          content:
            application/json:
              schema:
                properties:
                  postcode:
                    type: string
                    description: Postcode
                required:
                  - postcode
        responses:
          '200':
            description: Object with zoopla data
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    property:
                      $ref: ./zoopla.json#/definitions/zoopla
          '401':
            description: Not authorized
...